Use BigDecimal instead of Double
Because QueryDsl Kotlin serializer doesn’t yet support double
see https://github.com/querydsl/querydsl/pull/3346

+++ src/main/kotlin/griffio/planets/Planet.kt
@@ -1,10 +1,12 @@
 package griffio.planets
 
+import java.math.BigDecimal
+
 interface Planet {
 
     fun names(): String
 
-    fun au(): Double
+    fun au(): BigDecimal
 
     fun asString(): String = "${names()} is ${au()} au"
 }
